Consolidated validator package
