+++ src/main/java/com/rackspace/salus/monitor_management/web/model/telegraf/SqlServer.java
@@ -20,7 +20,7 @@ import com.rackspace.salus.monitor_management.web.model.ApplicableAgentType;
 import com.rackspace.salus.monitor_management.web.model.ApplicableMonitorType;
 import com.rackspace.salus.monitor_management.web.model.LocalPlugin;
 import com.rackspace.salus.monitor_management.web.model.SummaryField;
-import com.rackspace.salus.monitor_management.web.model.validator.ValidLocalHost;
+import com.rackspace.salus.monitor_management.web.validator.ValidLocalHost;
 import com.rackspace.salus.telemetry.model.AgentType;
 import com.rackspace.salus.telemetry.model.MonitorType;
 import java.util.List;
