+++ src/test/java/com/rackspace/salus/monitor_management/web/model/validator/ValidLocalHostTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Rackspace US, Inc.
+ * Copyright 2020 Rackspace US, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertThat;
 
+import com.rackspace.salus.monitor_management.web.validator.ValidLocalHost;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
